+++ station/lib/di/station_module_factory.dart
@@ -4,6 +4,7 @@ import 'package:station/repository/marker_repository.dart';
 import 'package:station/repository/open_time_repository.dart';
 import 'package:station/repository/origin_repository.dart';
 import 'package:station/repository/price_repository.dart';
+import 'package:station/repository/price_snapshot_repository.dart';
 import 'package:station/repository/station_repository.dart';
 import 'package:core/di/core_module_factory.dart';
 
@@ -20,6 +21,10 @@ class StationModuleFactory {
 
   static PriceRepository createPriceRepository() {
     return TanksteWebPriceRepository(createConfigRepository());
+  }
+
+  static PriceSnapshotRepository createPriceSnapshotRepository() {
+    return TanksteWebPriceSnapshotRepository(createConfigRepository());
   }
 
   static OpenTimeRepository createOpenTimeRepository() {