+++ CHANGELOG.md
@@ -9,6 +9,9 @@ CHANGELOG
 
 ## Upcoming version ##
 
+- 🧹 Dynamically load price fuel type
+- 🐞 Fix translation link
+- 🌟 Show price history chart on station details
 - 🌟 Add Danish translation
 - 🌟 Add Danish kroner currency